Report all missing path parameters, not just the first
This makes it possible to take remedial action such as copying the
parameters from the parent.

The method signature for build_request_path has been simplified. The
path argument was confusing and only used in one place to force the
collection path so a :collection option has been implemented instead.

This new options hash also implements :remove_used, which is useful
for build requests where you want to avoid parameters in the path
being duplicated in the query string.