Control whether associations are nested in parent saves using autosave
This new association option is similar to the option of the same name
found in ActiveRecord. When true, associated resources are always
nested in a parent save. When false, they are never nested. When nil,
only new resources are nested. Unlike ActiveRecord, the default is
true (rather than nil) in order to preserve existing behaviour.

This is also subject to whether send_only_modified_attributes is true
or not. When true, only modified resources will be sent, even if
autosave is true. Unlike before, it won't do strange things like fetch
an existing has_one resource when you assign a new one because of
change tracking.

4 of the specs currently fail because they require the fixes in my
fix-change-tracking branch.