Always treat id and custom primary keys as proper attributes
Not doing do was leading to change tracking not being applied in some
circumstances.

With a custom primary key, id is now treated as a formal alias. This
alters the behaviour slightly but setting an attribute with id= and
then getting a totally different value back with id is just as useless
as it is confusing. Some tests have had to be adjusted accordingly.

Note that you cannot set the primary key back to :id if it was set to
something else in an ancestor class. Undoing attribute aliases leads
to infinite loops.